connect/ca: split CA initialization logic between oss/enterprise
